Refactor space page code
