+++ client/src/pages/MySpaces/DeleteSpacesPage.tsx
