+++ client/src/pages/MySpaces/NewSpacePage.tsx
