+++ client/src/pages/MySpaces/index.ts
@@ -1,1 +1,3 @@
 export * from "./MySpacesPage";
+export * from "./NewSpacePage";
+export * from "./DeleteSpacesPage";
