+++ client/src/pages/SpaceDetail/Activity/index.ts
@@ -1,0 +1,2 @@
+export * from "./Transaction";
+export * from "./TransactionDetail";
