+++ client/src/pages/SpaceDetail/index.ts
@@ -1,2 +1,4 @@
+export * from "./SpaceDetail";
+export * from "./Activity";
 export * from "./NewTransactionPage";
 export * from "./DeleteTransactionPage";
