+++ client/src/pages/index.ts
@@ -1,5 +1,4 @@
 export * from "./HomePage";
 export * from "./auth";
 export * from "./MySpaces";
-export * from "./spaces";
-export * from "./transactions";
+export * from "./SpaceDetail";
