--- client/src/pages/spaces/index.ts
@@ -1,3 +1,0 @@
-export * from "./NewSpacePage";
-export * from "./DeleteSpacesPage";
-export * from "./SpacePage";
